// frontend/src/global/backend/apiV1/schema.d.ts
export interface components {
    schemas: {
        RsDataVoid: {
            resultCode: string;
            msg: string;
            data: unknown;
        };
        PostCommentModifyReqBody: {
            content: string;
        };
        PostModifyReqBody: {
            title: string;
            content: string;
        };
        PostWriteReqBody: {
            title: string;
            content: string;
        };
        PostWithAuthorDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            /** Format: int64 */
            authorId: number;
            authorName: string;
            title: string;
            content: string;
        };
        RsDataPostWithAuthorDto: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["PostWithAuthorDto"];
        };
        PostCommentWriteReqBody: {
            content: string;
        };
        PostCommentDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            authorName: string;
            content: string;
        };
        RsDataPostCommentDto: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["PostCommentDto"];
        };
        MemberJoinReqBody: {
            username: string;
            password: string;
            nickname: string;
        };
        MemberDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            nickname: string;
            isAdmin: boolean;
            profileImageUrl: string;
        };
        RsDataMemberDto: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["MemberDto"];
        };
        MemberLoginReqBody: {
            username: string;
            password: string;
        };
        MemberLoginResBody: {
            item: components["schemas"]["MemberDto"];
            apiKey: string;
            accessToken: string;
        };
        RsDataMemberLoginResBody: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["MemberLoginResBody"];
        };
        AdmPostCountResBody: {
            /** Format: int64 */
            all: number;
        };
        MemberWithUsernameDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            nickname: string;
            username: string;
            isAdmin: boolean;
        };
        PostDto: {
            /** Format: int64 */
            id: number;
            /** Format: date-time */
            createDate: string;
            /** Format: date-time */
            modifyDate: string;
            title: string;
            content: string;
        };
        RsDataPostDto: {
            resultCode: string;
            msg: string;
            data: components["schemas"]["PostDto"];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}